requirements.txt
Flask==1.1.2
PyGithub==1.51
zappa==0.51.0